requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
xgboost>=2.1
shap>=0.46
optuna>=4.1
matplotlib>=3.9
seaborn>=0.13
joblib>=1.4
langgraph>=0.4
langchain-core>=0.3
langchain-ollama>=1.1
langchain-groq>=1.0
langchain-openai>=0.3
langchain-anthropic>=0.3
fastapi>=0.115
uvicorn>=0.34
httpx>=0.28
supabase==2.3.4
sentence-transformers==2.3.1
pypdf==4.0.1
python-dotenv==1.0.0
groq==0.11.0
torch==2.2.0
